LLMService/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.25.0
chromadb==0.5.23
httpx==0.28.1
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.4
